+++ src/main/java/com/clova/anifriends/domain/recruitment/Recruitment.java
@@ -3,9 +3,9 @@ package com.clova.anifriends.domain.recruitment;
 import com.clova.anifriends.domain.applicant.Applicant;
 import com.clova.anifriends.domain.common.BaseTimeEntity;
 import com.clova.anifriends.domain.recruitment.exception.RecruitmentBadRequestException;
-import com.clova.anifriends.domain.recruitment.wrapper.RecruitmentContent;
-import com.clova.anifriends.domain.recruitment.wrapper.RecruitmentInfo;
-import com.clova.anifriends.domain.recruitment.wrapper.RecruitmentTitle;
+import com.clova.anifriends.domain.recruitment.vo.RecruitmentContent;
+import com.clova.anifriends.domain.recruitment.vo.RecruitmentInfo;
+import com.clova.anifriends.domain.recruitment.vo.RecruitmentTitle;
 import com.clova.anifriends.domain.shelter.Shelter;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
